⚡️ fix modal function #28
