+++ src/app/(home)/@modal/page.tsx
@@ -1,0 +1,3 @@
+export default async function Add() {
+  return null;
+}
