+++ src/app/(home)/default.ts
@@ -1,0 +1,1 @@
+export { default } from "./page";
