+++ src/app/(home)/friends/page.tsx
@@ -14,8 +14,7 @@ import {
 import dayjs from "dayjs";
 import { CheckSquare2 } from "lucide-react";
 
-import { getFolloweeAllPosts, getUserLatestPosts } from "@/lib/db/post";
-import { getProjectTitles } from "@/lib/db/project";
+import { getFolloweeAllPosts } from "@/lib/db/post";
 
 import Avatar from "../_components/avatar";
 import CraeteButton from "../_components/create-button";
@@ -32,26 +31,11 @@ export default async function Home() {
 
   const followeesAllPosts = await getFolloweeAllPosts(user.id);
 
-  const [projectTitles, latestPost] = await Promise.all([
-    getProjectTitles(user.id),
-    getUserLatestPosts(user.id),
-  ]);
-
-  const projectTitleArr = projectTitles.map((project) => project.title);
-
-  const isPostedToday =
-    latestPost?.created_at.getDate() !== new Date().getDate();
   return (
     <Box my={20}>
-      {!latestPost ? (
-        <NoPostCard>
-          <CraeteButton type={"text"} projects={projectTitleArr} />
-        </NoPostCard>
-      ) : isPostedToday ? (
-        <NoPostCard>
-          <CraeteButton type={"text"} projects={projectTitleArr} />
-        </NoPostCard>
-      ) : null}
+      <NoPostCard>
+        <CraeteButton type={"text"} />
+      </NoPostCard>
 
       {followeesAllPosts.map((post) => (
         <Card
